Exclusively use QueryOperators, even for pull operations
